refactor:  state로 hover 관리 삭제
